requirements.txt
# FastAPI関連
fastapi==0.115.14
uvicorn==0.35.0

# 設定管理
pydantic==2.11.7
pydantic-settings==2.10.1
typing-inspection==0.4.1

# Twitter API関連
tweepy==4.16.0

# OpenAI API関連
openai==1.93.3

# HTTP リクエスト
requests==2.32.4

# テスト関連
pytest==8.4.1
httpx==0.28.1

# 環境変数管理
python-dotenv==1.1.1 

# Redis
redis==6.2.0

# データベース関連
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# 暗号化関連
cryptography==44.0.0

# セキュリティ関連
slowapi==0.1.9
limits==5.5.0